deletecompleted tasks now uses the filter method
