+++ src/populate.js
@@ -2,6 +2,7 @@ import dragIcon from './images/drag_icon.svg';
 import deleteIcon from './images/delete.svg';
 import toggleCompleted from './complete-task.js';
 import deleteCompletedTasks from './delete-completed-tasks.js';
+import deleteAllComplete from './modules/delete-complete-tasks';
 
 const list = document.getElementById('list');
 
@@ -100,8 +101,7 @@ class TaskList extends Array {
   }
 
   deleteCompletedTasks() {
-    deleteCompletedTasks(this);
-    this.saveTasksToLocalStorage();
+    deleteAllComplete(this);
     this.render();
   }
 }
